fix markdown headings: cleanup only lines not starting with a doxygen comment
Lines containing two adjacent equal signs used to be removed even if
proper doxygen comment (starting with --!).  If the line with doxygen
comment contained two adjacent dashes, it used to be trimmed.

This makes impossible to use mardown headings level 1 and 2
(======,-------) and possibly breaks other things too.

This patch avoids this by cleaning only those lines, that do not start
with a doxygen comment.